refactor: turn individual filters into components
